Hover Operation
In the original version, when a the cursor hovered over a photo post, a
dark rectangular box appeared (stating how many likes the post had). But
this meant one of two things would happen if a user clicked on a photo
post. If the cursor was on the box, then they would be taken to the
permalink for that post. But if they clicked on the circle, yet not on
the box, then they would be taken to a blank page with nothing but the
photo (in other words where the picture was being hosted on tumblr's
server).
This would also happen on a mobile device, as the user could would be
forced to select a photo post by way of simply selecting the circle of
their liking, not knowing the consequences.

Therefore, I had to increase the size of the selection box so that when
a user hovered over a photo post, the entire circle would become opaque
& only allow one action from a click, that being the permalink (a normal
individual page for the post). Once on the permalink page, the user can
see the caption, which other tumblr users have hearted the photo, and
click on the photo itself to be taken to a link my choosing (typically
the portfolio/flickr of the photographer).

Also, the text stating the number of likes, was edited. I made the text
shade much lighter, and shifted its position to the top of the circle,
in essence to get it out of the way of the photo and making it nearly
invisible, but leaving the option in case any particular photo gains a
large following (which did occur).