CSS3 Hover: opacity, transform. More Caption Edits
Wanted user to feel which post they were about to select, so I took
advantage of the hover protocol and made the photo circles both darken
and increase in size (by 9%) when the cursor was above them.

Plus, I wanted the images below posts with multiple likes (represented
by user profiles in small circles) to be larger. Then, for the
corresponding font to align to the left, become bigger, and to have a
dark pink color for some contrast.

Note the [lack of] use of RGBa with my background settings, at a later
time I might want to experiment with adding shades of blue for the
subsequent hover effect.