Extensive Hover renovation/animation, and classes
The previous hover operation extended into the permalink page, therefore
the image and caption would enlarge when the user was on their unique
page, a very annoying effect. So first I over rode the transition in
several of the parent classes, including the .post class.

With the .permalink:hover class, I decided to animate (in a pulse
fashion) the opacity, and color it in a shade similar to the head logo.
Regarding the about page, I created three new classes. One to handle the
text, another for a giphytv (which is still under works) and another for
stylized text I have not yet utilized.