About Page
Essentially all of the code for the about page I had to write by hand.
The embed code for the vimeo and giphy were relatively simple to
implement. As for the youtube, I had to do extensive research on how to
embed it on mute. All in all it required a fair bit of javascript, but
it turned out well.  (I also made the vid autoplay, begin at a specific
point, and the progress bar white, via YoutubeAPI)

Inserted br tags to regulate spacing, and placed embeded vids in p tags.