TV Page
Just for fun I decided to add a page with an embed of giphy tv. The
probelm I was having before, and was solved thanks to some help directly
from GiphyLabs, was that the script tag did not have the
type="text/javascript" identifier. I've identified the folks over there
that the original embed prompt does not have this fix, but perhaps that
is more of an end-developer issue to resolve..

Then I wanted to have music to accompany the gifs. This meant finding a
music player that could embed and have an autoplay capability. Pandora
and Rdio were out of the running due to limited selections and
registration walls. Bandcamp didn't have autoplay. Soudcloud was close,
but it has Firefox compatibility issues. Therefore, I settled with
Jamendo. Their embed player has nice visuals, autoplay functionality,
and the site itself supports artists/creative commons works.